UIRenderer/UIRendererGroup improvements
Fixed a bug where text wouldn't render correctly if it was in a UI group. Made some general improvements to the UIRenderer and UIRenderer group code such as more comments. Renamed GLText2 to GLText, replacing the legacy version. GLText now requires a UIRenderer on construction. This is because it is not clear how to use GLText without a UIRenderer.