Beginning Re-make of SceneGame
The class DemoMap is a remake of SceneGame, it aims to render the map more efficiently by not using RendererModels. The new map rendering uses a texture atlas in which tiles are shared in one texture file. With the remake comes the goal to use the newer and more efficient layout practices that where a result of the SceneSwitcher update (they have already been done in the SceneMenuRemake).